requirements.txt
FormEncode==1.2.6
IMDbPY==5.0
SQLAlchemy==0.9.7
SQLObject==1.6.0
Tempita==0.5.2
WTForms==2.0.1
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
decorator==3.4.0
lxml==3.3.5
pbr==0.10.0
pymongo==2.7.2
rottentomatoes==1.1
six==1.7.3
sqlalchemy-migrate==0.9.1
tornado==4.0
wsgiref==0.1.2
wtforms-tornado==0.0.1
